+++ pkg/analyzer/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.33.6
+* Deprecated `AstNode.getAncestor` and introduced
+  `AstNode.thisOrAncestorMatching` as its replacement.
+
 ## 0.33.5
 * Add AnalysisSession.getResolvedLibrary()/ByElement() APIs.